sqrtf instead of sqrt

+++ src/hit/hit_sphere.c
@@ -6,7 +6,7 @@
 /*   By: jaberkro <jaberkro@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/06 15:46:19 by jaberkro      #+#    #+#                 */
-/*   Updated: 2022/11/18 16:40:10 by jaberkro      ########   odam.nl         */
+/*   Updated: 2022/12/01 12:49:40 by jaberkro      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ float	calculate_c(t_point change, float radius)
 {
 	float	result;
 
-	result = dot_points(change, change) - pow(radius, 2);
+	result = dot_points(change, change) - powf(radius, 2);
 	return (result);
 }
 
@@ -35,10 +35,10 @@ float	calculate_root(t_sphere *sphere, t_ray ray, float t_min, float t_max)
 	d = half_b * half_b - a * calculate_c(change, sphere->radius);
 	if (d < 0.0)
 		return (-1.0);
-	root = (-half_b - sqrt(d)) / a;
+	root = (-half_b - sqrtf(d)) / a;
 	if (root < t_min || t_max < root)
 	{
-		root = (-half_b + sqrt(d)) / a;
+		root = (-half_b + sqrtf(d)) / a;
 		if (root < t_min || t_max < root)
 			return (-1.0);
 	}
